remove App and Function enums; use string directly instead (#6)
this is to remove the use of enum for Apps and Functions and instead use
string directly.
So it's less confusing for users and more consistent when we keep adding
new apps and functions to our database (otherwise the sdk enums and
database deviated).

If in the future we want to introduce enums in some way, we can still
implement it.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Streamlined descriptions and examples to provide clearer guidance on
using app and function features.
  
- **Refactor**
- Updated internal reference methods in example scenarios to direct
users towards the platform’s current list of supported functionalities.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->